Redirect to root_url after register with invitation and refactor overrides to decorators
